# perf: improve rmul/rdiv/rsqrt operations