--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -35,6 +35,7 @@ jobs:
     - run: cargo clippy --version
     - run: cargo clippy --features i64
     - run: cargo clippy --all-targets --features i64
+    - run: cargo clippy --all-targets --features i64,i128
     - run: cargo clippy --all-targets --all-features
 
   test:
@@ -44,12 +45,12 @@ jobs:
     - run: rustc --version
     - run: cargo test --features i64
     - run: cargo test --features i128
-    - run: cargo test --no-default-features --lib --features i64
-    - run: cargo test --no-default-features --lib --features i128
-    - run: cargo test --no-default-features --lib --features std,i64
-    - run: cargo test --no-default-features --lib --features serde,i64
-    - run: cargo test --no-default-features --lib --features i64,parity
-    - run: cargo test --no-default-features --lib --features i128,parity
+    - run: cargo test --no-default-features --lib --test it --features i64
+    - run: cargo test --no-default-features --lib --test it --features i128
+    - run: cargo test --no-default-features --lib --test it --features std,i64
+    - run: cargo test --no-default-features --lib --test it --features serde,i64
+    - run: cargo test --no-default-features --lib --test it --features i64,parity
+    - run: cargo test --no-default-features --lib --test it --features i128,parity
     - run: cargo test --all-features
 
   run-example: