@loyd 2024-12-29
Need to update and check after https://github.com/Alexhuszagh/i256-rs/pull/40